Update greeting controllers and models
