+++ app/controllers/greetings_controller.rb
@@ -1,0 +1,6 @@
+class GreetingsController < ApplicationController
+  def random_greeting
+    greeting = Greeting.all.sample
+    render json: { greeting: }
+  end
+end
