+++ app/models/greeting.rb
@@ -1,0 +1,2 @@
+class Greeting < ApplicationRecord
+end
