+++ config/routes.rb
@@ -1,3 +1,3 @@
 Rails.application.routes.draw do
-  get '/random_greeting', to: 'greetings_api#random_greeting'
+  get '/random_greeting', to: 'greetings#random_greeting'
 end
