+++ db/seeds.rb
@@ -1,7 +1,11 @@
-# This file should contain all the record creation needed to seed the database with its default values.
-# The data can then be loaded with the bin/rails db:seed command (or created alongside the database with db:setup).
-#
-# Examples:
-#
-#   movies = Movie.create([{ name: "Star Wars" }, { name: "Lord of the Rings" }])
-#   Character.create(name: "Luke", movie: movies.first)
+greetings = [
+  'Hello, how are you?',
+  'Hi there! Welcome!',
+  'Greetings and salutations!',
+  'Hey, nice to see you!',
+  'Hola! ¿Cómo estás?'
+]
+
+greetings.each do |content|
+  Greeting.create(content: content)
+end
