+++ test/models/greeting_test.rb
@@ -1,0 +1,7 @@
+require "test_helper"
+
+class GreetingTest < ActiveSupport::TestCase
+  # test "the truth" do
+  #   assert true
+  # end
+end
